refactor: Major project cleanup - eliminate redundancy and improve organization
- Move redundant files to legacy with comprehensive documentation
- Create focused quantum_state_visualizer.py with extracted Wigner functions
- Eliminate duplicate implementations across src/examples/, src/utils/, src/quantum/

Legacy moves:
• src/examples/train_coordinate_quantum_gan_enhanced.py → legacy/enhanced_versions/
• src/utils/quantum_visualization_manager.py → legacy/complex_visualizers/
• src/utils/metrics.py → legacy/basic_metrics/
• src/quantum/core/sf_tutorial_circuit.py → legacy/quantum_circuits/

New additions:
• src/utils/quantum_state_visualizer.py - focused Wigner function visualization

Benefits:
- Eliminated all code redundancy
- Clear separation of concerns
- Preserved all valuable functionality
- Comprehensive legacy documentation
- Clean, maintainable architecture

All legacy moves documented in legacy/README.md with migration guide.
